+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Extend AliveTest and ScannerType enums.
+  [#235](https://github.com/greenbone/python-gvm/pull/235)
+
 ### Changed
 
 ### Fixed